formatting & bugfixes

+++ 4.Password/1.Network/thc-hydra/thc-hydra-update.sh
@@ -3,6 +3,26 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=git clone https://github.com/vanhauser-thc/thc-hydra.git
+GITREPOROOT=/opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc/thc-hydra
+GITCONFDIR=/opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc/thc-hydra/.git
+GITCLONEDIR=cd /opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/1.Network
+DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/1.Network
+DSKTPFL1=hydra-gtk.desktop
+DSKTPFL2=hydra.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  _____ _   _  ____      _   ___   ______  ____      _    
 |_   _| | | |/ ___|    | | | \ \ / /  _ \|  _ \    / \   
@@ -10,55 +30,55 @@ echo "${bold}
   | | |  _  | |__|_____|  _  | | | | |_| |  _ <  / ___ \ 
   |_| |_| |_|\____|    |_| |_| |_| |____/|_| \_\/_/   \_\
               
+UPDATE
 ${normal}"
 
-
-GITREPOROOT=/opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc/thc-hydra
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/1.Network
-DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/1.Network
-DSKTPFL1=hydra-gtk.desktop
-DSKTPFL2=hydra.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-
-mkdir -p /opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc
-cd /opt/ITSEC/4.Password/1.Network/thc-hydra/vanhauser-thc
-git clone https://github.com/vanhauser-thc/thc-hydra.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
 sudo make uninstall
 make clean
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
 
 ./configure
 make -j 4
 sudo make install
+
+mkdir -p $DSKTPFLSDEST
 rm -f $DSKTPFLSDEST/$DSKTPFL1
 rm -f $DSKTPFLSDEST/$DSKTPFL2
-
 cp $DSKTPFLS/$DSKTPFL1 $DSKTPFLSDEST/$DSKTPFL1
 cp $DSKTPFLS/$DSKTPFL2 $DSKTPFLSDEST/$DSKTPFL2
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
