+++ 4.Password/2.Local/johnny/johnny-update.sh
@@ -3,6 +3,27 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=git clone https://github.com/shinnok/johnny.git
+GITREPOROOT=/opt/ITSEC/4.Password/2.Local/johnny/shinnok/johnny
+GITCONFDIR=/opt/ITSEC/4.Password/2.Local/johnny/shinnok/johnny/.git
+GITCLONEDIR=/opt/ITSEC/4.Password/2.Local/johnny/shinnok
+EXECUTEABLE1=johnny
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/2.Local
+DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/2.Local
+DSKTPFL=johnny.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
      _  ___  _   _ _   _ _   ___   __
     | |/ _ \| | | | \ | | \ | \ \ / /
@@ -10,54 +31,55 @@ echo "${bold}
 | |_| | |_| |  _  | |\  | |\  | | |  
  \___/ \___/|_| |_|_| \_|_| \_| |_|  
        
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/4.Password/2.Local/johnny/shinnok/johnny
-GITREPOGITFILE=$GITREPOROOT/.git
-EXECUTEABLE1=johnny
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/2.Local
-DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/2.Local
-DSKTPFL=johnny.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/4.Password/2.Local/johnny/shinnok
-cd /opt/ITSEC/4.Password/2.Local/johnny/shinnok
-git clone https://github.com/shinnok/johnny.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
+    
 cd $GITREPOROOT
-sudo rm -r /usr/local/bin/johnny
-sudo updatedb
-sudo ldconfig
-git clean -f
-git pull
-git submodule init
-git submodule update --recursive
-#
+GITRESET
+GITSBMDLINIT
+
 export QT_SELECT=qt5
 qmake
 make -j$(nproc)
 
-sudo ln -s $GITREPOROOT/$EXECUTEABLE1 /usr/local/bin/$EXECUTEABLE1
+sudo rm -f $BINDIR/$EXECUTEABLE1
+sudo ln -s $GITREPOROOT/$EXECUTEABLE1 $BINDIR/$EXECUTEABLE1
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
+
