+++ 4.Password/2.Local/pyrit/pyrit-update.sh
@@ -3,6 +3,24 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/JPaulMora/Pyrit.git
+GITREPOROOT=/opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora/Pyrit
+GITCONFDIR=/opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora/Pyrit/.git
+GITCLONEDIR=/opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/2.Local
+DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/2.Local
+DSKTPFL=pyrit.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
 echo "${bold}
  ______   ______  ___ _____ 
 |  _ \ \ / /  _ \|_ _|_   _|
@@ -10,47 +28,48 @@ echo "${bold}
 |  __/ | | |  _ < | |  | |  
 |_|    |_| |_| \_\___| |_|  
             
+UPDATE
 ${normal}"
-
-GITREPOROOT=/opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora/Pyrit
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/4.Password/2.Local
-DSKTPFLSDEST=/home/$USER/.local/share/applications/4.Password/2.Local
-DSKTPFL=pyrit.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora
-cd /opt/ITSEC/4.Password/2.Local/pyrit/JPaulMora
-git clone https://github.com/JPaulMora/Pyrit.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
 
 cd $GITREPOROOT
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
 
 sudo python setup.py install
+
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
