+++ 5.Database/1.SQL/bbqsql/##bbqsql-update.sh
@@ -3,6 +3,25 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/Neohapsis/bbqsql.git
+GITCONFDIR=/opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis/bbqsql/.git
+GITREPOROOT=/opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis/bbqsql
+GITCLONEDIR=/opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/5.Database/1.SQL
+DSKTPFLSDEST=/home/$USER/.local/share/applications/5.Database/1.SQL
+DSKTPFL=bbqsql.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/master
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ____  ____   ___  ____   ___  _     
 | __ )| __ ) / _ \/ ___| / _ \| |    
@@ -10,50 +29,52 @@ echo "${bold}
 | |_) | |_) | |_| |___) | |_| | |___ 
 |____/|____/ \__\_\____/ \__\_\_____|
           
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis/bbqsql
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/5.Database/1.SQL
-DSKTPFLSDEST=/home/$USER/.local/share/applications/5.Database/1.SQL
-DSKTPFL=bbqsql.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis
-cd /opt/ITSEC/5.Database/1.SQL/bbqsql/Neohapsis
-git clone https://github.com/Neohapsis/bbqsql.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-sudo rm -f /usr/local/bin/bbqsql
-#
+if git checkout master &&
+    git fetch origin master &&
+    [ `git rev-list HEAD...origin/master --count` != 0 ] &&
+    git merge origin/master
+then
 
 cd $GITREPOROOT
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
+
+sudo rm -f /usr/local/bin/bbqsql
 
 sudo -H python setup.py install
+
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST 
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
